requirements.txt
streamlit>=1.36.0
openai>=1.42.0
python-dotenv>=1.0.1
pydantic>=2.8.2
jinja2>=3.1.4
python-docx>=1.1.2
